requirements.txt
beautifulsoup4==4.7.1
dash==1.6.1
dash-core-components==1.5.1
dash-daq==0.1.0
dash-html-components==1.0.2
dash-renderer==1.2.1
dash-table==4.5.1
dask==1.1.4
datashader==0.7.0
Flask==1.0.2
Flask-Compress==1.4.0
folium==0.10.0
gunicorn==20.0.3
IMDbPY==6.9.dev20191124
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==1.0.0
jupyterlab-server==0.2.0
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
pandas==0.24.2
pandocfilters==1.4.2
plotly==4.3.0
wordcloud==1.6.0.post1+g8217e20
